backend/requirements.txt
# Backend Django - Tatlight
# Python 3.10+

# Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# JWT Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9  # Pour PostgreSQL (production)
# ou pour dev local avec SQLite, pas besoin

# Image handling
Pillow==10.2.0

# Environment variables
python-decouple==3.8

# Password validation
argon2-cffi==23.1.0

# Paiements
stripe==8.2.0
# flutterwave-python==1.1.0  # À installer si besoin

# Utils
python-dateutil==2.8.2
pytz==2024.1

# Dev dependencies
django-debug-toolbar==4.2.0